Correction du code
Correction des id en int
Quelques renommages
Ajouts de quelques mappers supplémentaires
Réarrgement de resetDatabase

+++ src/main/java/com/excilys/mviegas/speccdb/controlers/IEditorComputerControler.java
@@ -56,7 +56,7 @@ public interface IEditorComputerControler {
 	/**
 	 * @return Entreprise fabricant l'ordinateur
 	 */
-	int getIdCompany();
+	long getIdCompany();
 
 	/**
 	 * @param pIdCompany Fixe l'entreprise fabricant l'ordinateur
